Reset retryCount and lastError after task instance successfully retries
For looping tasks (e.g. those with `while(true) { ... }`), if the task is retried successfully, the next time it fails, it will only be retried if there are available delays left in the policy (in the context of `DelayPolicy`-based policies). This seems counter-intuitive, so it should probably be changed so that successfully retrying the task clears the slate. There may be some instances where that wouldn't be the desired behavior, but that seems outside of the typical use-case and should be implemented in a custom policy.